# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "summary": "Register a new user",
        "description": "Creates a new user with email, username and password. Returns user id, email and username.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              },
              "examples": {
                "default": {
                  "summary": "New user",
                  "value": {
                    "email": "user@example.com",
                    "username": "johndoe",
                    "password": "secret123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Email or username already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Log in",
        "description": "Authenticates with email and password. Returns a JWT access token for Authorization header.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "examples": {
                "default": {
                  "summary": "Credentials",
                  "value": {
                    "email": "user@example.com",
                    "password": "secret123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "Current user",
        "description": "Returns the authenticated user (requires valid JWT in Authorization header).",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (missing or invalid token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/search": {
      "get": {
        "operationId": "GamesController_search",
        "summary": "Search games",
        "description": "Searches the local catalog by name. If no results, tries external provider (e.g. BGG) and persists results. externalAvailable indicates if external was used.",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query (game name)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSearchResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ]
      }
    },
    "/games/{id}": {
      "get": {
        "operationId": "GamesController_getOne",
        "summary": "Get game by ID",
        "description": "Returns a single game by UUID. Returns 404 if not found.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Game UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ]
      }
    },
    "/games": {
      "post": {
        "operationId": "GamesController_create",
        "summary": "Create game (trusted users)",
        "description": "Creates a new game in the catalog. Requires JWT and trusted-user role.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameDto"
              },
              "examples": {
                "default": {
                  "summary": "New game",
                  "value": {
                    "name": "Catan",
                    "year": 1995,
                    "minPlayers": 3,
                    "maxPlayers": 4,
                    "playTime": 90
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Game created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not a trusted user)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/search": {
      "get": {
        "operationId": "SearchController_search",
        "summary": "Global search",
        "description": "Search both games and users by query. When authenticated, excludes current user from results and includes follow flags.",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Games and users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSearchResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Search"
        ]
      }
    },
...
  },
  "info": {
    "title": "BG Hub API",
    "description": "Board game collection, wishlist, play logs and social features.",
    "version": "0.0.2",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "format": "email",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "description": "Unique username",
            "example": "johndoe"
          },
          "password": {
            "type": "string",
            "description": "Password (min 8 characters)",
            "example": "secret123",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "username",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User UUID",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "description": "Unique username",
            "example": "johndoe"
          }
        },
        "required": [
          "id",
          "email",
          "username"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password",
            "example": "secret123",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token for Authorization header",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "GameDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Game UUID",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Game name",
            "example": "Catan"
          },
          "externalId": {
            "type": "string",
            "description": "External provider ID (e.g. BGG)",
            "example": "13"
          },
          "apiRef": {
            "type": "string",
            "description": "External API reference",
            "example": "bgg"
          },
          "imageUrl": {
            "type": "object",
            "description": "Cover image URL",
            "nullable": true
          },
          "year": {
            "type": "object",
            "description": "Year published",
            "nullable": true,
            "example": 1995
          },
          "minPlayers": {
            "type": "object",
            "description": "Minimum number of players",
            "nullable": true,
            "example": 3
          },
          "maxPlayers": {
            "type": "object",
            "description": "Maximum number of players",
            "nullable": true,
            "example": 4
          },
          "playTime": {
            "type": "object",
            "description": "Playing time in minutes",
            "nullable": true,
            "example": 90
          },
          "complexityWeight": {
            "type": "object",
            "description": "Complexity/weight (e.g. BGG)",
            "nullable": true
          },
          "categories": {
            "description": "Category names",
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mechanics": {
            "description": "Mechanic names",
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "object",
            "description": "Game description",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "externalId",
          "apiRef",
          "imageUrl",
          "year",
          "minPlayers",
          "maxPlayers",
          "playTime",
          "complexityWeight",
          "categories",
          "mechanics",
          "description"
        ]
      },
      "GameSearchResponseDto": {
        "type": "object",
        "properties": {
          "games": {
            "description": "Matching games",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameDto"
            }
          },
          "externalAvailable": {
            "type": "boolean",
            "description": "Whether external provider (e.g. BGG) was available for this search",
            "example": true
          }
        },
        "required": [
          "games",
          "externalAvailable"
        ]
      },
      "CreateGameDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Game name",
            "example": "Catan"
          },
          "year": {
            "type": "number",
            "description": "Year published",
            "example": 1995,
            "minimum": 1
          },
          "minPlayers": {
            "type": "number",
            "description": "Minimum number of players",
            "minimum": 1
          },
          "maxPlayers": {
            "type": "number",
            "description": "Maximum number of players",
            "minimum": 1
          },
          "playTime": {
            "type": "number",
            "description": "Playing time in minutes",
            "minimum": 1
          },
          "complexityWeight": {
            "type": "number",
            "description": "Weight/complexity (e.g. BGG weight)",
            "minimum": 0
          },
          "categories": {
            "description": "Category names",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mechanics": {
            "description": "Mechanic names",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "description": "Game description"
          },
          "imageUrl": {
            "type": "string",
            "description": "Cover image URL"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "SearchUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User UUID",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "johndoe"
          },
          "displayName": {
            "type": "object",
            "description": "Display name",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "description": "Avatar URL",
            "nullable": true
          },
          "followsYou": {
            "type": "boolean",
            "description": "True if this user follows the current user"
          },
          "isFollowing": {
            "type": "boolean",
            "description": "True if the current user follows this user"
          }
        },
        "required": [
          "id",
          "username",
          "displayName",
          "avatarUrl",
          "followsYou",
          "isFollowing"
        ]
      },
      "GlobalSearchResponseDto": {
        "type": "object",
        "properties": {
          "games": {
            "description": "Matching games",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameDto"
            }
          },
          "users": {
            "description": "Matching users (excludes current user when authenticated)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchUserDto"
            }
          }
        },
        "required": [
          "games",
          "users"
        ]
      },
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code",
            "example": 404
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Resource not found"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      }
    }
  }
}